Moved the DeleteSubLiveData.php command to the merge folder to make it a bit more clear about what its used for. Fixed the account delete. Made the cooperation argument.
